[IMP] (website_)survey: hide header/footer when taking a certification
Purpose
=======

We need to hide header and footer of survey and website layouts when
the user is taking a certification ("certificate" field of survey.survey).

This is done to have a more "immersive" taking and avoid that the user
clicks on other links in the layout header or footer and moves to another page.

closes odoo/odoo#31503

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>