[FIX] stock: useless kwargs on ml action_done
commit a3a305e61e6cc5aed72adcc855a7dac5b184fefc introduced a kwarg on
picking and stock move in order to check if the validation should run a
backorder creation or not. However the validation of stock move line is
a level lower and the concept of backorder does not exist at this level.
So the kwarg in action_done is an error and should be removed.