[REV] models: do not erase master version
This commit has been backported in 11.0 [1] and then reverted [2],
because it introduced a regression in web_studio, discovered at opw-1947380

As no forward-port has been done between the backport and the revert,
the forward-port of the revert also contains the patch itself, resulting
in a noop.

Actually revert 60fbed853479feefa10a9ae442354adf11f49ab9 in branch `12.0`.

[1] 1b62529685f1973711641e6e978941bf898485dd
[2] a5fb5e5e13f9042394927158259f149279766fc4

closes odoo/odoo#31637

Signed-off-by: Christophe Simonis <chs@odoo.com>