[IMP] purchase: keep purchase representative empty in the purchase order
purpose of this task is do not set default purchase representative in the purchase order when the PO is created automatically.
When the purchase order is created automatically, this purchase order would be assigned to odoobot.
Odoobot is not a real user. so nobody notified any activity of that purchase order.
so keep the responsible field empty when the PO is created automatically.
Purchase manager can easily filter "unassigned" purchase order and he can assign unassigned PO to someone (real user).

Task-1940489

closes odoo/odoo#31729

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>